BC-8093 - rename `src/infra/logging` to `src/infra/logger`
