+++ src/infra/metrics/metrics.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { LoggerModule } from '../logging/logger.module.js';
+import { LoggerModule } from '../logger/logger.module.js';
 import { MetricsController } from './api/metrics.controller.js';
 import { MetricsService } from './metrics.service.js';
