+++ src/infra/redis/redis.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ConfigurationModule } from '../configuration/configuration.module.js';
-import { LoggerModule } from '../logging/logger.module.js';
+import { LoggerModule } from '../logger/logger.module.js';
 import { RedisConfig } from './redis.config.js';
 import { RedisService } from './redis.service.js';
